Moves rheo_packages into the packages/ module
Turns rheo_packages.rs into crates/core/src/packages/mod.rs, matching
the concern-folder pattern, and renames the module to packages.
Updates the crate::packages::RheoPackages references.
