+++ crates/core/src/lib.rs
@@ -3,10 +3,10 @@ pub mod build;
 pub mod compile;
 pub mod config;
 pub mod diagnostics;
+pub mod packages;
 pub mod parser;
 pub mod plugins;
 pub mod reticulate;
-pub mod rheo_packages;
 pub mod templates;
 pub mod util;
 pub mod world;
