+++ crates/core/src/plugins/typst_manifest.rs
@@ -1,7 +1,7 @@
 use crate::config::PluginAssets;
+use crate::packages::RheoPackages;
 use crate::parser::extract_package_imports;
 use crate::plugins::{PackageAssets, ResolvedPackage, parse_package_spec};
-use crate::rheo_packages::RheoPackages;
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
